fix(REGISTRY-0000): Don't show NotificationModal when it doesn't have new-old pairs. (#833)
