+++ src/organisms/NotificationsMenu/hooks/useGetNotifications.tsx
@@ -1,8 +1,8 @@
 import { useInfiniteQuery } from "@tanstack/react-query";
 import getNotificationsQuery from "@/services/notifications/getNotificationsQuery";
 
-const useGetNotifcations = (userId: number) => {
+const useGetNotifications = (userId: number) => {
   return useInfiniteQuery(getNotificationsQuery(userId));
 };
 
-export default useGetNotifcations;
+export default useGetNotifications;
